MRG - Merged revision from b22-release branch starting from last revision of merge: 11090-11105, except website related changes in html files.
git-svn-id: http://scarab.tigris.org/svn/scarab@11131 3c7165bb-d2ef-0310-b724-bafd9a86081f

+++ trunk/src/java/org/tigris/scarab/om/Issue.java
@@ -3122,7 +3122,7 @@ public class Issue
             .getInstance(NUMBERKEY_0);
         ActivityManager
             .createTextActivity(newIssue, zeroAttribute, activitySet2,
-                                ActivityType.ISSUE_MOVED,
+                                ActivityType.ISSUE_COPIED,
                                 getUniqueId(), newIssue.getUniqueId());
 
         newIssue.index();
@@ -3395,16 +3395,20 @@ public class Issue
             for (Iterator it = dependencies.iterator(); it.hasNext(); )
             {
                 Depend depend = (Depend)it.next();
-                ActivitySet deleteSet = 
-                    this.doDeleteDependency(activitySet, depend, user);
-                for (Iterator act = deleteSet.getActivityList().iterator(); act.hasNext(); )
-                {
-                    activitySet.addActivity((Activity)act.next());
-                }
-                NotificationManagerFactory.getInstance()
-                        .addActivityNotification(ActivityType.ISSUE_DELETED,
-                                activitySet, this, user);
+                if(!depend.getDeleted()){
+                	ActivitySet deleteSet = 
+                        this.doDeleteDependency(activitySet, depend, user);
+                    for (Iterator act = deleteSet.getActivityList().iterator(); act.hasNext(); )
+                    {
+                        activitySet.addActivity((Activity)act.next());
+                    }
+                }        
             }
+            
+            NotificationManagerFactory.getInstance()
+            .addActivityNotification(ActivityType.ISSUE_DELETED,
+                    activitySet, this, user);
+            
             save();
         }
     }
